✨ [feat] InfoLinkView 구현 #29
